Phase_5/requirements.txt
torch
torchvision
onnx
onnxruntime
opencv-python
flask
gunicorn
pyyaml
tqdm
numpy
Pillow
scikit-image
